Add Testrun table output after the testrun is finished
