+++ pkg/testrunner/run.go
@@ -16,21 +16,12 @@ package testrunner
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/url"
-	"path"
-	"regexp"
-
-	"github.com/gardener/test-infra/pkg/testmachinery"
+	tmv1beta1 "github.com/gardener/test-infra/pkg/apis/testmachinery/v1beta1"
 	trerrors "github.com/gardener/test-infra/pkg/testrunner/error"
+	"github.com/gardener/test-infra/pkg/util"
 	"github.com/go-logr/logr"
 	"github.com/hashicorp/go-multierror"
-	"k8s.io/api/extensions/v1beta1"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	tmv1beta1 "github.com/gardener/test-infra/pkg/apis/testmachinery/v1beta1"
-	"github.com/gardener/test-infra/pkg/util"
 )
 
 // GetTestruns returns all testruns of a RunList as testrun array
@@ -104,58 +95,6 @@ func (r *Run) Exec(log logr.Logger, config *Config, prefix string) {
 		r.Testrun.Status.Phase = tmv1beta1.PhaseStatusTimeout
 		r.Error = trerrors.NewTimeoutError(fmt.Sprintf("maximum wait time of %d is exceeded by Testrun %s", config.Timeout, r.Testrun.GetName()))
 	}
-}
 
-func GetArgoURL(tmClient client.Client, tr *tmv1beta1.Testrun) (string, error) {
-	// get argo url from the argo ingress if possible
-	// return err if the ingress cannot be found
-	argoIngress := &v1beta1.Ingress{}
-	err := tmClient.Get(context.TODO(), client.ObjectKey{Name: "argo-ui", Namespace: "default"}, argoIngress)
-	if err != nil {
-		return "", err
-	}
-
-	if len(argoIngress.Spec.Rules) == 0 {
-		return "", errors.New("no rules defined")
-	}
-	rule := argoIngress.Spec.Rules[0]
-	if len(rule.HTTP.Paths) == 0 {
-		return "", errors.New("no http backend defined")
-	}
-
-	protocol := "http://"
-	if len(argoIngress.Spec.TLS) != 0 {
-		protocol = "https://"
-	}
-	argoUrl, err := url.ParseRequestURI(protocol + rule.Host)
-	if err != nil {
-		return "", err
-	}
-	argoUrl.Path = path.Join(argoUrl.Path, "workflows", tr.Namespace, testmachinery.GetWorkflowName(tr))
-
-	return argoUrl.String(), nil
-}
-
-// GetClusterDomainURL tries to derive the cluster domain url from an grafana ingress if possible. Returns an error if the ingress cannot be found or is in unexpected form.
-func GetClusterDomainURL(tmClient client.Client) (string, error) {
-	// try to derive the cluster domain url from grafana ingress if possible
-	// return err if the ingress cannot be found
-	if tmClient == nil {
-		return "", nil
-	}
-	ingress := &v1beta1.Ingress{}
-	err := tmClient.Get(context.TODO(), client.ObjectKey{Namespace: "monitoring", Name: "grafana"}, ingress)
-	if err != nil {
-		return "", fmt.Errorf("cannot get grafana ingress: %v", err)
-	}
-	if len(ingress.Spec.Rules) == 0 {
-		return "", fmt.Errorf("cannot get ingress rule from ingress %v", ingress)
-	}
-	host := ingress.Spec.Rules[0].Host
-	r, _ := regexp.Compile("[a-z]+\\.ingress\\.(.+)$")
-	matches := r.FindStringSubmatch(host)
-	if len(matches) < 2 {
-		return "", fmt.Errorf("cannot regex cluster domain from ingress %v", ingress)
-	}
-	return matches[1], nil
+	fmt.Println(RunList{r}.RenderTable())
 }
